requirements.txt
Django==5.1.12
django-environ==0.11.2
dj-database-url==2.2.0
psycopg2-binary==2.9.9
boto3==1.35.0
minio==7.2.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
strawberry-graphql-django>=0.47.2
gunicorn==23.0.0
daphne>=4
sphinx==7.1.2
myst-parser==2.0.0
sphinxcontrib-mermaid==0.9.2
furo==2024.1.29
requests
pytest
pytest-django
pytest-cov
pytest-socket
coverage
hypothesis
vulture
diff-cover
grimp
deptry
mutmut
ruff
blake3>=0.4.1
jsonschema>=4.0.0
jmespath>=1.0.0
PyYAML>=6.0
modal==1.1.4
websockets>=12.0
httpx>=0.24.0
